Module skeletal_model::bone

Describes skeletal bones and their relationships.

For more info about skeletal bones and the role they play in the skeleton, see the module-level documentation on crate::skeleton.

Modules

bone_kind 🔒

bone_map

Structs

BoneMap

Provides a map of BoneKind -> T. Every possible BoneKind must have a corresponding value.

Enums

BoneKind

Describes the various types of bones in the skeleton.