Enum skeletal_model::bone::bone_kind::BoneKind

#[repr(u8)]
pub enum BoneKind {

    Neck,
    Chest,
    Waist,
    Hip,
    ThighL,
    ThighR,
    AnkleL,
    AnkleR,
    FootL,
    FootR,
    UpperArmL,
    UpperArmR,
    ForearmL,
    ForearmR,
    WristL,
    WristR,

}

Describes the various types of bones in the skeleton.

All of this information is static rather than dynamic information. “Static” here refers to information that does not change - this information is “hard coded”. This includes:

The parent/child relationships between the various types of bones

The pose of the bones when performing a calibration

Names of the bones

Etc

Where possible, this information is provided as const functions so that they can be evaluated at compile-time.

There is also dynamic information associated with bones. “Dynamic” here refers to the fact that some bone data cannot be known up front and changes as the skeletal model receives inputs and produces outputs. This data is stored in the Skeleton as an Edge.

BoneKind is also represented as a u8, so it can be used as an index for an array. This is used for example in BoneMap. Please note that we make no stability guarantees for the particular value that any variant gets, only that these values are contiguous and start at 0. Use the variant directly or refer to the various functions implemented on this type for stability.

Variants

Neck

Chest

Waist

Hip

ThighL

ThighR

AnkleL

AnkleR

FootL

FootR

UpperArmL

UpperArmR

ForearmL

ForearmR

WristL

WristR

Implementations

impl BoneKind

pub const fn max() -> BoneKind

The bone with the largest integer value

pub const MAX: BoneKind = _

pub const fn min() -> BoneKind

The bone with the smallest integer value

pub const MIN: BoneKind = _

pub const fn root() -> Self

The root bone of the skeletal graph/tree.

pub const ROOT: BoneKind = _

pub const fn num_types() -> usize

Returns the number of unique kinds of bones. This is equivalent to the number of variants in BoneKind

pub const NUM_TYPES: usize = 16usize

pub const fn children(&self) -> &'static [Self]

Returns the children of any particular bone.

The slice is 'static, which means the lifetime of the returned slice lives for the entire duration of the program. This is because the parent/child relationship of bones is known at compile-time.

pub const fn parent(&self) -> Option<BoneKind>

The parent of a bone.

pub fn iter() -> Map<RangeInclusive<u8>, fn(_: u8) -> BoneKind>

pub fn calibration_rotation(self) -> Global<UnitQuat>

Returns the initial calibration pose of the bone. Rotating the up vector by this rotation would cause it to point in the same target direction as the bone.

pub fn calibration_rotation_local(self) -> Local<UnitQuat>

Returns the initial calibration pose of the bone, as a rotation relative to the parent bone. See also: Self::calibration_rotation