Struct skeletal_model::bone::BoneMap

pub struct BoneMap<T>([T; 16]);

Provides a map of BoneKind -> T. Every possible BoneKind must have a corresponding value.

This is more efficient than a HashMap, because it is done without allocation inline, and the BoneKinds don’t need to be hashed.

Example

let mut m = BoneMap::default();
for b in BoneKind::iter() {
    m[b] = format!("this is a {b:?}");
}

assert_eq!(m[BoneKind::Chest], format!("this is a {:?}", BoneKind::Chest))

Tuple Fields
0: [T; 16]
Implementations

impl<T> BoneMap<T>

pub fn new(map: [T; 16]) -> Self

pub fn iter(&self) -> Iter<'_, T>

Gets an iterator over the (key, value) pairs of the BoneMap.

Iteration is guaranteed to start at BoneKind::root() but beyond that, iteration order is not guaranteed. However, iteration is exhaustive over the various kinds of bones.

pub fn iter_mut(&mut self) -> IterMut<'_, T>

Gets a mutable iterator over the (key, value) pairs of the BoneMap.

See also: Self::iter()

pub fn map<U>(self, f: impl FnMut(BoneKind, T) -> U) -> BoneMap<U>

Applies a function to each element of the BoneMap, mapping it from T to U.