Struct skeletal_model::skeleton::Skeleton

pub struct Skeleton {
    bone_map: BoneMap<EdgeIndex>,
    graph: Dag<Node, Edge>,
}

The Skeleton provides a way of reading, writing, and solving for the pose of a human wearing FBT.

See the crate::skeleton module for more information.

Fields
bone_map: BoneMap<EdgeIndex>graph: Dag<Node, Edge>
Implementations

impl Skeleton

pub fn new(config: &SkeletonConfig) -> Self

Creates a new Skeleton from SkeletonConfig. Initially, the skeleton will not have any input trackers or output trackers.